requirements.txt
gunicorn
Flask
Flask-Login
Flask-SQLAlchemy
SQLAlchemy
Werkzeug
python-dotenv
Flask-Migrate